debug

+++ sources/bresenham_line_algo.c
@@ -6,7 +6,7 @@
 /*   By: tzanchi <tzanchi@student.42berlin.de>      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/25 16:37:59 by tzanchi           #+#    #+#             */
-/*   Updated: 2023/08/04 15:33:44 by tzanchi          ###   ########.fr       */
+/*   Updated: 2023/08/07 19:04:38 by tzanchi          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -47,7 +47,7 @@ void	horizontal(t_br_utils ut, mlx_image_t *img, int colour)
 	while (ut.x0 != ut.x1)
 	{
 		if (ut.x0 >= 0 && ut.x0 <= WIDTH && ut.y0 >= 0 && ut.y0 <= HEIGTH)
-			mlx_put_pixel(img, (int)ut.x0, (int)ut.y0, colour);
+			mlx_put_pixel(img, ut.x0, ut.y0, colour);
 		ut.x0 += ut.step_x;
 	}
 }
@@ -59,7 +59,7 @@ void	vertical(t_br_utils ut, mlx_image_t *img, int colour)
 	while (ut.y0 != ut.y1)
 	{
 		if (ut.x0 >= 0 && ut.x0 <= WIDTH && ut.y0 >= 0 && ut.y0 <= HEIGTH)
-			mlx_put_pixel(img, (int)ut.x0, (int)ut.y0, colour);
+			mlx_put_pixel(img, ut.x0, ut.y0, colour);
 		ut.y0 += ut.step_y;
 	}
 }
@@ -72,7 +72,7 @@ void	plot(t_vect_3 a, t_vect_3 b, mlx_image_t *img, int colour)
 
 	init_bresenham_utils(&ut, a, b);
 	if (ut.delta_x == 0 && ut.delta_y == 0)
-		mlx_put_pixel(img, (int)ut.x0, (int)ut.y0, colour);
+		mlx_put_pixel(img, ut.x0, ut.y0, colour);
 	else if (ut.delta_y == 0)
 		horizontal(ut, img, colour);
 	else if (ut.delta_x == 0)
@@ -82,7 +82,7 @@ void	plot(t_vect_3 a, t_vect_3 b, mlx_image_t *img, int colour)
 		while (ut.x0 != ut.x1)
 		{
 			if (ut.x0 >= 0 && ut.x0 < WIDTH && ut.y0 >= 0 && ut.y0 <= HEIGTH)
-				mlx_put_pixel(img, (int)ut.x0, (int)ut.y0, colour);
+				mlx_put_pixel(img, ut.x0, ut.y0, colour);
 			if (ut.p >= 0)
 			{
 				ut.y0 += ut.step_y;
